Fix remaining game-core LP64/clang warnings (clean x64 build)
A clean FreeBSD amd64 (clang) rebuild surfaced 5 warnings the incremental
build had hidden; all fixed properly (no suppression):

- char_item.cpp:2372,2388 - item->GetSocket() (TimeT64) printed with %lld;
  cast the arg to (long long) (same class as the db-core format fixes).
- trigger.cpp:64 - m_iMinDistance(~(1L << 31)) relied on 32-bit long: on
  LP64 1L<<31 is a 64-bit 0x80000000 whose complement truncates to int with
  a -Wconstant-conversion warning. The intent is INT_MAX; use it directly.
- questlua_dungeon.cpp:1287 - lua_pushnumber(L, LONG_MAX): LONG_MAX is
  2^63-1 on LP64 and cannot be represented exactly as a Lua double
  (-Wimplicit-const-int-float-conversion). The original 32-bit sentinel was
  INT_MAX (2^31-1, exact in double); use INT_MAX to preserve it.
- input_main.cpp guild-CG switch - GUILD_SUBHEADER_CG_DEPOSIT_MONEY /
  _WITHDRAW_MONEY were unhandled (-Wswitch). Handle them explicitly
  (consume+ignore, like the other unused sub-headers) instead of a default.

Game core rebuilds 0 warnings / 0 errors on FreeBSD amd64.

+++ Lead-Server-Source/game/src/char_item.cpp
@@ -2369,7 +2369,7 @@ bool CHARACTER::UseItemEx(LPITEM item, TItemPos DestCell)
 								{
 									if (item->GetVnum() == DRAGON_HEART_VNUM)
 									{
-										snprintf(buf, sizeof(buf), "Inc %ds by item{VN:%d SOC%d:%lld}", ret, item->GetVnum(), ITEM_SOCKET_CHARGING_AMOUNT_IDX, item->GetSocket(ITEM_SOCKET_CHARGING_AMOUNT_IDX));
+										snprintf(buf, sizeof(buf), "Inc %ds by item{VN:%d SOC%d:%lld}", ret, item->GetVnum(), ITEM_SOCKET_CHARGING_AMOUNT_IDX, (long long) item->GetSocket(ITEM_SOCKET_CHARGING_AMOUNT_IDX));
 									}
 									else
 									{
@@ -2385,7 +2385,7 @@ bool CHARACTER::UseItemEx(LPITEM item, TItemPos DestCell)
 								{
 									if (item->GetVnum() == DRAGON_HEART_VNUM)
 									{
-										snprintf(buf, sizeof(buf), "No change by item{VN:%d SOC%d:%lld}", item->GetVnum(), ITEM_SOCKET_CHARGING_AMOUNT_IDX, item->GetSocket(ITEM_SOCKET_CHARGING_AMOUNT_IDX));
+										snprintf(buf, sizeof(buf), "No change by item{VN:%d SOC%d:%lld}", item->GetVnum(), ITEM_SOCKET_CHARGING_AMOUNT_IDX, (long long) item->GetSocket(ITEM_SOCKET_CHARGING_AMOUNT_IDX));
 									}
 									else
 									{
